Pulled avalon editor and floating toolbar into a control, cleaning up DocumentView heaps
