+++ src/MarkPad.Services/MarkPadExtensions/IMarkPadExtension.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-namespace MarkPad.Services.MarkPadExtensions
+﻿namespace MarkPad.Services.MarkPadExtensions
 {
 	public interface IMarkPadExtension
 	{
