+++ src/MarkPad/Document/DocumentViewModel.cs
@@ -2,12 +2,10 @@ using System;
 using System.IO;
 using System.Net;
 using System.Threading.Tasks;
-using System.Windows.Media;
 using System.Windows.Threading;
 using Caliburn.Micro;
 using CookComputing.XmlRpc;
 using ICSharpCode.AvalonEdit.Document;
-using MarkPad.Framework;
 using MarkPad.HyperlinkEditor;
 using MarkPad.Services.Implementation;
 using MarkPad.Services.Interfaces;
@@ -22,7 +20,6 @@ namespace MarkPad.Document
         private static readonly ILog Log = LogManager.GetLog(typeof(DocumentViewModel));
 
         private readonly IDialogService dialogService;
-        private readonly ISettingsProvider settings;
         private readonly IWindowManager windowManager;
         private readonly ISiteContextGenerator siteContextGenerator;
         private readonly Func<string, IMetaWeblogService> getMetaWeblog;
@@ -34,10 +31,9 @@ namespace MarkPad.Document
         private string filename;
         private ISiteContext siteContext;
 
-        public DocumentViewModel(IDialogService dialogService, ISettingsProvider settings, IWindowManager windowManager, ISiteContextGenerator siteContextGenerator, Func<string, IMetaWeblogService> getMetaWeblog )
+        public DocumentViewModel(IDialogService dialogService, IWindowManager windowManager, ISiteContextGenerator siteContextGenerator, Func<string, IMetaWeblogService> getMetaWeblog )
         {
             this.dialogService = dialogService;
-            this.settings = settings;
             this.windowManager = windowManager;
             this.siteContextGenerator = siteContextGenerator;
             this.getMetaWeblog = getMetaWeblog;
