+++ src/MarkPad/Document/MarkdownEditor.xaml
@@ -1,0 +1,90 @@
+﻿<UserControl x:Class="MarkPad.Document.MarkdownEditor"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:avalonedit="http://icsharpcode.net/sharpdevelop/avalonedit"
+             xmlns:cal="http://www.caliburnproject.org"
+             xmlns:controls="clr-namespace:MarkPad.XAML.Controls;assembly=MarkPad.XAML"
+             xmlns:Document="clr-namespace:MarkPad.Document" mc:Ignorable="d" 
+             d:DesignHeight="300" d:DesignWidth="300"
+             Loaded="EditorLoaded">
+    <UserControl.Resources>
+        <Style TargetType="{x:Type avalonedit:TextArea}">
+            <Setter Property="FocusVisualStyle" Value="{x:Null}"/>
+            <Setter Property="SelectionBrush">
+                <Setter.Value>
+                    <SolidColorBrush Color="{DynamicResource {x:Static SystemColors.HighlightColorKey}}" Opacity="0.7"/>
+                </Setter.Value>
+            </Setter>
+            <Setter Property="SelectionBorder">
+                <Setter.Value>
+                    <Pen Brush="{DynamicResource {x:Static SystemColors.HighlightBrushKey}}" Thickness="1"/>
+                </Setter.Value>
+            </Setter>
+            <Setter Property="SelectionForeground" Value="{DynamicResource {x:Static SystemColors.HighlightTextBrushKey}}"/>
+            <Setter Property="Template">
+                <Setter.Value>
+                    <ControlTemplate TargetType="{x:Type avalonedit:TextArea}">
+                        <DockPanel Focusable="False">
+                            <ItemsControl DockPanel.Dock="Left"
+                                      Focusable="False"
+                                      Margin="0,0,5,0"
+                                          FontSize="10"
+                                          FontFamily="Segoe UI"
+                                      Padding="10,0,0,0"
+                                      Background="#adDEDEDE"
+                                      ItemsSource="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=LeftMargins}">
+                                <ItemsControl.ItemsPanel>
+                                    <ItemsPanelTemplate>
+                                        <StackPanel Orientation="Horizontal" />
+                                    </ItemsPanelTemplate>
+                                </ItemsControl.ItemsPanel>
+                            </ItemsControl>
+                            <ContentPresenter
+                            Panel.ZIndex="-1"
+                            Focusable="False"
+                            Content="{Binding RelativeSource={RelativeSource TemplatedParent}, Path=TextView}"/>
+                        </DockPanel>
+                    </ControlTemplate>
+                </Setter.Value>
+            </Setter>
+        </Style>
+    </UserControl.Resources>
+    <UserControl.InputBindings>
+        <KeyBinding Command="Document:FormattingCommands.ToggleBold" Key="B" Modifiers="Control" />
+        <KeyBinding Command="Document:FormattingCommands.ToggleItalic" Key="I" Modifiers="Control" />
+        <KeyBinding Command="Document:FormattingCommands.ToggleCode" Key="J" Modifiers="Control" />
+        <KeyBinding Command="Document:FormattingCommands.SetHyperlink" Key="K" Modifiers="Control" />
+    </UserControl.InputBindings>
+    <Grid>
+        <controls:FloatingToolBar x:Name="floatingToolBar" PlacementTarget="{Binding ElementName=Editor}" CommandTarget="{Binding ElementName=root}" VerticalOffset="-60" HorizontalOffset="-30">
+            <Border Background="#f0f0f0" Padding="3,2,3,2" BorderThickness="1" CornerRadius="1" BorderBrush="#999999">
+                <StackPanel Orientation="Horizontal">
+                    <Button Width="22"  Command="Document:FormattingCommands.ToggleBold" ToolTip="Bold">B</Button>
+                    <Button Width="22" Command="Document:FormattingCommands.ToggleItalic" ToolTip="Italic">I</Button>
+                    <Button Width="44" Command="Document:FormattingCommands.ToggleCode" ToolTip="Code" Margin="5,0,0,0">Code</Button>
+                    <Button Width="78" Command="Document:FormattingCommands.ToggleCodeBlock" ToolTip="Code block">Code block</Button>
+                    <Button Width="42" Command="Document:FormattingCommands.SetHyperlink" ToolTip="Hyperlink">www</Button>
+                </StackPanel>
+            </Border>
+        </controls:FloatingToolBar>
+        
+        <avalonedit:TextEditor
+            x:Name="Editor"
+            WordWrap="True"
+            ShowLineNumbers="True"
+            Document="{Binding Document}"
+            cal:Message.Attach="[Event TextChanged]=[Action Update]"
+            Grid.Column="0"
+            Margin="10,0,10,10" 
+            TextOptions.TextFormattingMode="Display"
+            VerticalAlignment="Stretch"
+            HorizontalAlignment="Stretch" Padding="0,0,10,0"
+            PreviewKeyDown="EditorPreviewKeyDown"            >
+            <avalonedit:TextEditor.Effect>
+                <DropShadowEffect BlurRadius="10" Color="Black" Opacity="0.25" Direction="270" />
+            </avalonedit:TextEditor.Effect>
+        </avalonedit:TextEditor>
+    </Grid>
+</UserControl>
