+++ src/MarkPad/MarkPadExtensions/IDocumentViewExtension.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using MarkPad.Document;
+﻿using MarkPad.Document;
 using MarkPad.Services.MarkPadExtensions;
 
 namespace MarkPad.MarkPadExtensions
