+++ src/MarkPad/MarkPadExtensions/SpellCheck/SpellCheckExtension.cs
@@ -6,34 +6,34 @@ using MarkPad.Services.Interfaces;
 
 namespace MarkPad.MarkPadExtensions.SpellCheck
 {
-	public class SpellCheckExtension : IDocumentViewExtension
-	{
-		public string Name { get { return "Spell check"; } }
+    public class SpellCheckExtension : IDocumentViewExtension
+    {
+        public string Name { get { return "Spell check"; } }
 
-		readonly ISpellingService _spellingService;
+        readonly ISpellingService spellingService;
 
-		IList<SpellCheckProvider> _providers = new List<SpellCheckProvider>();
+        readonly IList<SpellCheckProvider> providers = new List<SpellCheckProvider>();
 
-		public SpellCheckExtension(ISpellingService spellingService)
-		{
-			_spellingService = spellingService;
-		}
+        public SpellCheckExtension(ISpellingService spellingService)
+        {
+            this.spellingService = spellingService;
+        }
 
-		public void ConnectToDocumentView(DocumentView view)
-		{
-			if (_providers.Any(p => p.View == view)) throw new ArgumentException("View already has a spell check provider connected", "view");
-			
-			var provider = new SpellCheckProvider(_spellingService, view);
-			_providers.Add(provider);
-		}
+        public void ConnectToDocumentView(DocumentView view)
+        {
+            if (providers.Any(p => p.View == view)) throw new ArgumentException("View already has a spell check provider connected", "view");
 
-		public void DisconnectFromDocumentView(DocumentView view)
-		{
-			var provider = _providers.FirstOrDefault(p => p.View == view);
-			if (provider == null) return;
+            var provider = new SpellCheckProvider(spellingService, view);
+            providers.Add(provider);
+        }
 
-			provider.Disconnect();
-			_providers.Remove(provider);
-		}
-	}
+        public void DisconnectFromDocumentView(DocumentView view)
+        {
+            var provider = providers.FirstOrDefault(p => p.View == view);
+            if (provider == null) return;
+
+            provider.Disconnect();
+            providers.Remove(provider);
+        }
+    }
 }
